refactor: externalize constants and types (Phase 3)
Move all constants and types to dedicated files for better organization:
- Created constants/contract-networks.ts with SUPPORTED_CONTRACT_NETWORKS
- Consolidated all contract-related types in types.ts
- Updated all imports across components
- Fixed TypeScript readonly array compatibility with spread operator

This reduces LinkContractAddressButton from ~196 to ~148 lines (additional
24% reduction). Combined with all phases, total reduction is 64% (416 → 148).

All three refactoring phases now complete.
