+++ REFACTOR_PLAN_LinkContractAddressButton.md
@@ -48,18 +48,22 @@ The component has **~430 lines** with multiple responsibilities:
 
 **Result**: Component reduced from ~248 lines to ~196 lines (additional ~21% reduction)
 
-### **Phase 3: Configuration & Constants** (Low Impact)
+### **Phase 3: Configuration & Constants** ✅ COMPLETED (Low Impact)
 
-**3.1 Move `SUPPORTED_NETWORKS` to constants file**
-- Create `constants/contract-networks.ts`
-- Rename to SUPPORTED_CONTRACT_NETWORKS
-- Export as `SUPPORTED_NETWORKS` and optionally add network metadata
-- **Impact**: Remove ~45 lines, improve maintainability
+**3.1 Move `SUPPORTED_NETWORKS` to constants file** ✅
+- Created `constants/contract-networks.ts` with SUPPORTED_CONTRACT_NETWORKS
+- Added TypeScript const assertion for type safety
+- Exported SupportedContractNetwork type
+- **Impact**: 46-line constants file, better maintainability
 
-**3.2 Create types file for all interfaces**
-- Move `LinkContractAddressesButtonProps` to separate types file
-- Co-locate with `NetworkAddressPair` and `InvalidInfo`
-- **Impact**: Better organization
+**3.2 Consolidate types in types file** ✅
+- Moved `LinkContractAddressesButtonProps` to types file
+- Moved `InvalidInfo` to types file
+- Moved `ContractAddressItemProps` to types file
+- All contract-related types now co-located
+- **Impact**: 32-line types file, improved organization
+
+**Result**: Component reduced from ~196 lines to ~148 lines (additional ~24% reduction)
 
 ## Proposed File Structure After Refactoring
 
@@ -83,16 +87,27 @@ utilities/
   └── contractKey.ts                         (existing)
 ```
 
-## Current Outcome (After Phase 1 & 2)
-- **LinkContractAddressButton**: ~416 lines → ~196 lines (53% reduction)
+## Final Outcome ✅ ALL PHASES COMPLETE
+
+### Component Size Reduction
+- **Original**: ~416 lines
+- **After Phase 1**: ~248 lines (40% reduction)
+- **After Phase 2**: ~196 lines (53% reduction)
+- **After Phase 3**: ~148 lines (64% total reduction)
+
+### Created Files
+- **3 Custom Hooks**: useContractAddressPairs, useContractAddressValidation, useContractAddressSave
+- **2 UI Components**: ContractAddressDialog, ContractAddressList
+- **1 Constants File**: contract-networks.ts
+- **1 Types File**: Enhanced types.ts with all contract-related types
+
+### Benefits Achieved
 - **Improved testability**: Each hook can be tested independently
-- **Better separation of concerns**: UI, state, and business logic separated
+- **Better separation of concerns**: UI, state, and business logic fully separated
 - **Increased reusability**: Dialog and list components reusable across the app
 - **Easier maintenance**: Smaller, focused modules
-
-## Final Expected Outcome (After Phase 3)
-- **LinkContractAddressButton**: ~196 lines → ~150 lines (64% total reduction from original)
-- All constants and types externalized for better organization
+- **Type safety**: Centralized types with TypeScript const assertions
+- **Better organization**: Constants and types externalized
 
 ## Recommended Execution Order
 1. Start with **Phase 1.1** (useContractPairs) - lowest risk, immediate benefit
